requirements.txt
python-dotenv
tqdm
opencv-python
pytube
google-api-python-client
google-cloud-aiplatform